closes #30, closes #29, closes #17, closes #3, closes #1; Frontend changes. Backend fixes

+++ StudentPerfomance-Backend/SPerfomance.Application.Students.Module/Repository/StudentCommandRepository.cs
@@ -9,40 +9,68 @@ using SPerfomance.Domain.Module.Shared.Entities.StudentGroups;
 using SPerfomance.Domain.Module.Shared.Entities.StudentGroups.Errors;
 using SPerfomance.Domain.Module.Shared.Entities.Students;
 using SPerfomance.Domain.Module.Shared.Entities.Students.Errors;
+using SPerfomance.Domain.Module.Shared.Extensions;
 
 namespace SPerfomance.Application.Students.Module.Repository;
 
 internal sealed class StudentCommandRepository
 {
 	private readonly ApplicationDb _context = new ApplicationDb();
+
 	public async Task Commit() => await _context.SaveChangesAsync();
+
 	public async Task<Result<Student>> Create(StudentSchema student, IRepositoryExpression<StudentGroup> getGroup, IRepositoryExpression<Student> findDulicate)
 	{
 		if (await _context.Students.AnyAsync(findDulicate.Build()))
 			return Result.Failure<Student>(new StudentDublicateRecordBookError(student.Recordbook).ToString());
-		StudentGroup? group = await _context.Groups.FirstOrDefaultAsync(getGroup.Build());
+
+		StudentGroup? group = await _context.Groups
+		.Include(g => g.EducationPlan)
+		.ThenInclude(p => p.Direction)
+		.FirstOrDefaultAsync(getGroup.Build());
+
 		if (group == null)
 			return Result.Failure<Student>(new GroupNotFoundError().ToString());
+
 		Student entry = student.CreateDomainObject(group);
 		Result create = group.AddStudent(entry);
-		if (create.IsFailure) return Result.Failure<Student>(create.Error);
+
+		if (create.IsFailure)
+			return Result.Failure<Student>(create.Error);
+
 		entry.SetNumber(await GenerateEntityNumber());
 		_context.Groups.Attach(entry.Group);
+
 		await _context.Students.AddAsync(entry);
 		await Commit();
 		return entry;
 	}
-	public async Task<Result<Student>> Remove(IRepositoryExpression<Student> getStudent)
+
+	public async Task<Result<Student>> Remove(IRepositoryExpression<Student> getStudent, IRepositoryExpression<StudentGroup> getGroup)
 	{
+		StudentGroup? group = await _context.Groups
+		.Include(g => g.EducationPlan)
+		.ThenInclude(p => p.Direction)
+		.Include(g => g.Students)
+		.FirstOrDefaultAsync();
+		if (group == null)
+			return Result.Failure<Student>(new GroupNotFoundError().ToString());
+
 		Student? student = await _context.Students.FirstOrDefaultAsync(getStudent.Build());
 		if (student == null)
 			return Result.Failure<Student>(new StudentNotFoundError().ToString());
+
+		Result remove = group.RemoveStudent(student);
+		if (remove.IsFailure)
+			return Result.Failure<Student>(remove.Error);
+
 		await _context.Students.Where(s => s.Id == student.Id).ExecuteDeleteAsync();
 		return student;
 	}
+
 	public async Task<int> GenerateEntityNumber()
 	{
-		int count = await _context.Students.CountAsync();
-		return count + 1;
+		int[] numbers = await _context.Students.Select(s => s.EntityNumber).ToArrayAsync();
+		return numbers.GetOrderedValue();
 	}
 }
